+++ Src/CSharpier.VSCode/src/CSharpierProcessProvider.ts
@@ -2,7 +2,6 @@ import { Disposable, Extension, TextEditor, window, workspace } from "vscode";
 import { Logger } from "./Logger";
 import * as path from "path";
 import * as semver from "semver";
-import { execSync } from "child_process";
 import * as convert from "xml-js";
 import { ICSharpierProcess, NullCSharpierProcess } from "./CSharpierProcess";
 import { CSharpierProcessSingleFile } from "./CSharpierProcessSingleFile";
@@ -10,6 +9,7 @@ import { CSharpierProcessPipeMultipleFiles } from "./CSharpierProcessPipeMultipl
 import * as fs from "fs";
 import { InstallerService } from "./InstallerService";
 import { CustomPathInstaller } from "./CustomPathInstaller";
+import { execDotNet } from "./DotNetProvider";
 
 export class CSharpierProcessProvider implements Disposable {
     warnedForOldVersion = false;
@@ -138,16 +138,13 @@ export class CSharpierProcessProvider implements Disposable {
         let outputFromCsharpier: string;
 
         try {
-            outputFromCsharpier = execSync(`dotnet csharpier --version`, {
-                cwd: directoryThatContainsFile,
-                env: { ...process.env, DOTNET_NOLOGO: "1" },
-            })
+            outputFromCsharpier = execDotNet(`csharpier --version`, directoryThatContainsFile)
                 .toString()
                 .trim();
 
             this.logger.debug(`dotnet csharpier --version output: ${outputFromCsharpier}`);
-            const versionWithoutHash = outputFromCsharpier.split("+")[0]
-            this.logger.debug(`Using ${versionWithoutHash} as the version number.`)
+            const versionWithoutHash = outputFromCsharpier.split("+")[0];
+            this.logger.debug(`Using ${versionWithoutHash} as the version number.`);
             return versionWithoutHash;
         } catch (error: any) {
             const message = !error.stderr ? error.toString() : error.stderr.toString();
@@ -212,7 +209,7 @@ export class CSharpierProcessProvider implements Disposable {
             }
 
             if (!this.customPathInstaller.ensureVersionInstalled(version)) {
-                this.logger.debug(`Unable to validate install of version ${version}`)
+                this.logger.debug(`Unable to validate install of version ${version}`);
                 this.displayFailureMessage();
                 return NullCSharpierProcess.instance;
             }
@@ -230,7 +227,11 @@ export class CSharpierProcessProvider implements Disposable {
                 }
                 return new CSharpierProcessSingleFile(this.logger, customPath);
             } else {
-                const csharpierProcess = new CSharpierProcessPipeMultipleFiles(this.logger, customPath, directory);
+                const csharpierProcess = new CSharpierProcessPipeMultipleFiles(
+                    this.logger,
+                    customPath,
+                    directory,
+                );
                 if (csharpierProcess.processFailedToStart) {
                     this.displayFailureMessage();
                 }