+++ Src/CSharpier.VSCode/src/CustomPathInstaller.ts
@@ -1,8 +1,9 @@
 import { Logger } from "./Logger";
 import * as path from "path";
 import * as fs from "fs";
-import { execSync } from "child_process";
 import { workspace } from "vscode";
+import { getDotNetRoot, execDotNet } from "./DotNetProvider";
+import { execSync } from "child_process";
 
 export class CustomPathInstaller {
     logger: Logger;
@@ -38,7 +39,7 @@ export class CustomPathInstaller {
 
         const command = `dotnet tool install csharpier --version ${version} --tool-path "${pathToDirectoryForVersion}"`;
         this.logger.debug("Running " + command);
-        execSync(command);
+        execDotNet(command);
 
         return this.validateInstall(pathToDirectoryForVersion, version);
     }
@@ -46,26 +47,30 @@ export class CustomPathInstaller {
     private validateInstall(pathToDirectoryForVersion: string, version: string): boolean {
         try {
             const output = execSync(`"${this.getPathForVersion(version)}" --version`, {
-                env: { ...process.env, DOTNET_NOLOGO: "1" },
+                env: { ...process.env, DOTNET_ROOT: getDotNetRoot() },
             })
                 .toString()
                 .trim();
 
             this.logger.debug(`"${this.getPathForVersion(version)}" --version output: ${output}`);
-            const versionWithoutHash = output.split("+")[0]
-            this.logger.debug(`Using ${versionWithoutHash} as the version number.`)
+            const versionWithoutHash = output.split("+")[0];
+            this.logger.debug(`Using ${versionWithoutHash} as the version number.`);
 
             if (versionWithoutHash === version) {
                 this.logger.debug("CSharpier at " + pathToDirectoryForVersion + " already exists");
                 return true;
-            }
-            else {
-                this.logger.warn("Version of " + versionWithoutHash + " did not match expected version of " + version)
+            } else {
+                this.logger.warn(
+                    "Version of " +
+                        versionWithoutHash +
+                        " did not match expected version of " +
+                        version,
+                );
             }
         } catch (error: any) {
             const message = !error.stderr ? error.toString() : error.stderr.toString();
             this.logger.warn(
-                "Exception while running 'dotnet csharpier --version' in " +
+                "Exception while running 'dotnet-csharpier --version' in " +
                     pathToDirectoryForVersion,
                 message,
             );